refactor: colocando UserContext na pasta hook
