+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Economigos",
+  "name": "Economigos",
   "icons": [
     {
       "src": "favicon.ico",
